Configuracion de controller de code y command
