+++ server/graphql/resolvers/mutation/codesMutation.js
@@ -1,13 +1,7 @@
-import codesSchema from '../../../models/codesSchema'
+import codesController from '../../../controllers/codesController'
 
 export default {
   addCode (_, {title, description, code}) {
-    return codesSchema.create({title, description, code}, (err, data) => {
-      if (err) {
-        return err
-      } else {
-        return data
-      }
-    })
+    return codesController.create(title, description, code)
   }
 }
