+++ server/graphql/resolvers/query/commandsQuery.js
@@ -1,21 +1,10 @@
-import commandsSchema from '../../../models/commandsSchema'
-var mongoose = require('mongoose')
+import commandsController from '../../../controllers/commandsControllers'
 
 export default {
   commands () {
-    console.log('llego aqui ? ')
-    return commandsSchema.find({}).exec() // 58f2a89a2463a62024505ecd
+    return commandsController.all()
   },
   command (root, arg, context) {
-    if (arg.id && mongoose.Types.ObjectId.isValid(arg.id)) {
-      return commandsSchema.findById(arg.id)
-        .then((data) => {
-          return data
-        }).catch((error) => {
-          return error
-        })
-    } else {
-      return new Error('Id invalid')
-    }
+    return commandsController.findById(root, arg, context)
   }
 }
