+++ server/routers/commandsPrivateRouter.js
@@ -3,7 +3,7 @@
 import express from 'express'
 import passport from 'passport'
 import commandsSchema from '../models/commandsSchema'
-import CommandPrivateControllers from '../controllers/commandPrivateControllers'
+/*import CommandPrivateControllers from '../controllers/commandPrivateControllers'
 
 const router = express.Router()
 const commandPrivateControllers = new CommandPrivateControllers(commandsSchema)
@@ -18,5 +18,5 @@ router.get('/command', passport.authenticate('bearer', {session: false}), comple
 router.put('/command/:idCommand', passport.authenticate('bearer', {session: false}), completeUser, commandPrivateControllers.update)
 router.delete('/command/:idCommand', passport.authenticate('bearer', {session: false}), completeUser, commandPrivateControllers.delete)
 router.get('/command/:idCommand', passport.authenticate('bearer', {session: false}), completeUser, commandPrivateControllers.detail)
-
-export default router
+*/
+// export default router
